### 2017-03-31 #219 @whedon
[PRE REVIEW]: filltex: Automatic queries to ADS and INSPIRE databases to fill LaTex bibliography
**Submitting author:** @dgerosa (<a href="http://orcid.org/0000-0002-0933-3579">Davide Gerosa</a>)
**Repository:** <a href="https://github.com/dgerosa/filltex" target ="_blank">https://github.com/dgerosa/filltex</a>
**Version:** v1.0.0
