### 2017-04-10 dgerosa/filltex#1 @lnielsen
Authorship of fillbib.py
I'm reviewing openjournals/joss-reviews#222, and you have correctly given credit for the [fillbib.py](https://github.com/dgerosa/filltex/blob/master/fillbib.py) script to Michele Vallisneri. The script is  largely the same with some improvements. However, what is not clear to me what license the original script has?

All I'm able to find is a [Copyright Michele Vallisneri](http://www.vallis.org/salon/summary-2.html) in the bottom of the page which in practice means that you can't  take the script and include it in your code (the similarity between the two are too big to claim that your script is written from scratch).

Now, I think likely Michele Vallisneri have no problem with the script being included here, so it's probably just a matter of getting him to explicitly state the license for the script or that he places the script in the public domain.

Also, the original script I think should include a copyright statement from Michele Vallisneri.
