### 2017-04-10 dgerosa/filltex#2 @lnielsen
Better detection of ADS and arXiv identifiers
The [``fillbib.py``](https://github.com/dgerosa/filltex/blob/master/fillbib.py#L60) script can improve the detection of ADS and arXiv identifiers.

You can use the https://idutils.readthedocs.io package for that or simply copy over the corresponding [regular expressions](https://github.com/inveniosoftware/idutils/blob/master/idutils/__init__.py#L44-L59) if you don't want the dependency.
