### 2017-04-10 dgerosa/filltex#2 @lnielsen
Related to openjournals/joss-reviews#222
