### 2017-04-10 dgerosa/filltex#3 @lnielsen
Python 3 support
It would be relatively easy to add support for Python 3 and would make the script more future proof.

Helpful things that would make it work on Python 3 is:

- Add ``from __future__ import print_statement`` and rewrite ``print "test"`` to ``print("test")``
- Reading files will by default give you back byte strings in Python 3 so you might need to detect string type (six library is your friend, specifically [these lines](https://bitbucket.org/gutworth/six/src/92e1c746e1abfbdb94705b821bc93766083efc54/six.py?at=default&fileviewer=file-view-default#six.py-40:53))

openjournals/joss-reviews#222
