### 2017-04-10 dgerosa/filltex#4 @lnielsen
Unit tests
The repo includes a test file which is great, but it serves more the purpose of an example of how it works. The package ought to include some minimal unit tests that validates the actual output with expected output.

https://docs.pytest.org/en/latest/ can likely help you here.

Related to openjournals/joss-reviews#222

A bonus would be to hook it up with Travis-CI which shouldn't take long.
