### 2017-04-10 dgerosa/filltex#5 @lnielsen
Reporting bugs
I would recommend that bugs are reported to the GitHub issue tracking instead of to the email address (dgerosa@caltech.edu). This ensures that the issues are public so you can get help from others or people don't double report the same issue. Also, we all move organisations from time to time so better keep the source code and issues together the same place.

Related to openjournals/joss-reviews#222
