### 2017-04-10 dgerosa/filltex#6 @lnielsen
Consider making it into a Python package
The repository could be made into Python package and be released on Python Package Index.

The installation could be reduced to:
```
$ pip install filltex
$ filltex install-texshop
$ filltex <file>
```

You could use the [click library](http://click.pocoo.org/5/) for handling the CLI (and install the command line tool). This also makes unit testing the CLI pretty easy.

You could easily use dependencies such as ``requests`` and ``idutils``  libraries to do part of the work.

You could refactor the script into API + CLI. The API being citation extraction from .aux files etc (might be there's a Python library already doing this).

You would need modifying the PATH environment.

You would need to deal properly rewrite the filltex bash script into Python or call it from Python.

If this is something you'd like I'll happily provide more details.

openjournals/joss-reviews#222
