### 2017-04-10 dgerosa/filltex#7 @lnielsen
Using cURL from Python script
Use of requests library was removed from the original script and instead using ``curl``. It's fine not to have the requests library dependency, but calling curl from within Python seems a bit strange to me as it needs to spawn a new process. I think it would be better to use the standard library's https://docs.python.org/2/library/urllib.html. That said, requests is a really good library and it solves many of the issues you'll have with urllib.

openjournals/joss-reviews#222
