### 2017-04-10 dgerosa/filltex#8 @lnielsen
Add guidelines for how to contribute to the package and how to seek help
You are already saying how to report bugs, so perhaps just add a sentence that contributions are welcome and you can open an issue if you need help.

openjournals/joss-reviews#222
