### 2017-04-10 #222 @lnielsen
I've had an initial review of the repository. I like that it's a small simple script to automate the fetching the citations automatically from ADS/INSPIRE. I think it's definitely something that can save researchers time when writing so that's much appreciated.

I think the following issues should be resolved before the paper can be accepted:

- [ ] dgerosa/filltex#1
- [ ] dgerosa/filltex#2
- [ ] dgerosa/filltex#4
- [ ] dgerosa/filltex#5
- [ ] dgerosa/filltex#8

These optional issues will help improve the quality of the software but are not necessary in order to accept the paper:

- [ ] dgerosa/filltex#3
- [ ] dgerosa/filltex#6
