### 2017-04-10 dgerosa/filltex#5 @dgerosa
Thanks for this comment. I have updated the readme accordingly.
