### 2017-04-10 dgerosa/filltex#7 @dgerosa
I have updated the code to use urllib as suggested. I would like to avoid the requests library
 dependency.
