### 2017-04-10 dgerosa/filltex#2 @dgerosa
Thanks for this comment.

- The code now simply guesses if the requested reference is in ADS or INSPIRE by reading its first character. This is based on the simple observation that ADS citations start with numbers (year) while INSPIRE citations stat with letter (author). It seems to me a solid classification, as it looks very unlikely that any of the two databases will change this in the future.
- While keeping its functionalities, I believe the code as is it now is simpler and more readable, because it explicitly lists the urls which are scraped.
- I would not like to include citations directly from the arxiv, as I believe it is important people cite the published version of a paper whenever available.
