+++ fillbib.py
@@ -12,83 +12,104 @@ python fillbib.py <tex_file> <bib_file>. If <bib_file> is absent, it will try to
 '''
 import sys, os, re, commands, urllib
 
-if len(sys.argv)==2:     # Get the name of the bibfile from the aux file
-    basename = sys.argv[1].split('.tex')[0]
-    auxfile = basename + '.aux'
+def ads_url(c): # url for ADS citations
+    return
+
+def ads_citation(c): # download single ADS citation
+
+    f= urllib.urlopen("http://adsabs.harvard.edu/cgi-bin/nph-bib_query?bibcode="+c+"&data_type=BIBTEX&db_key=AST&nocookieset=1")
+    bib = f.read()
+    f.close()
+    bib = "@"+bib.split("@")[1]
+
+    if 'arXiv' in c: # Take care of preprint on ADS
+        bib = bib.split("{")[0]+"{"+c+","+",".join(bib.split(",")[1:])
+
+    return bib
+
+def inspire_citation(c): # download single INSPIRE citation
+
+    f= urllib.urlopen("https://inspirehep.net/search?p="+c+"&of=hx&em=B&sf=year&so=d&rg=1")
+    bib = f.read()
+    f.close()
+    bib = "@"+bib.split("@")[1].split('</pre>')[-2]
+
+    return bib
+
+def test_ads(): # test single ADS web scraping (both published articles and preprints)
+    test_key = ["2016PhRvL.116f1102A","2016arXiv160203837T"]
+    known_output= '@ARTICLE{2016PhRvL.116f1102A,\n   author = {{Abbott}, B.~P. and {Abbott}, R. and {Abbott}, T.~D. and {Abernathy}, M.~R. and \n\t{Acernese}, F. and {Ackley}, K. and {Adams}, C. and {Adams}, T. and \n\t{Addesso}, P. and {Adhikari}, R.~X. and et al.},\n    title = "{Observation of Gravitational Waves from a Binary Black Hole Merger}",\n  journal = {Physical Review Letters},\narchivePrefix = "arXiv",\n   eprint = {1602.03837},\n primaryClass = "gr-qc",\n     year = 2016,\n    month = feb,\n   volume = 116,\n   number = 6,\n      eid = {061102},\n    pages = {061102},\n      doi = {10.1103/PhysRevLett.116.061102},\n   adsurl = {http://adsabs.harvard.edu/abs/2016PhRvL.116f1102A},\n  adsnote = {Provided by the SAO/NASA Astrophysics Data System}\n}\n\n'
+    assert [ads_citation(tk) == known_output for tk in test_key]
+
+def test_inspire(): # test single INSPIRE web scraping
+    test_key = "Abbott:2016blz"
+    known_output= '@article{Abbott:2016blz,\n      author         = "Abbott, B. P. and others",\n      title          = "{Observation of Gravitational Waves from a Binary Black\n                        Hole Merger}",\n      collaboration  = "Virgo, LIGO Scientific",\n      journal        = "Phys. Rev. Lett.",\n      volume         = "116",\n      year           = "2016",\n      number         = "6",\n      pages          = "061102",\n      doi            = "10.1103/PhysRevLett.116.061102",\n      eprint         = "1602.03837",\n      archivePrefix  = "arXiv",\n      primaryClass   = "gr-qc",\n      reportNumber   = "LIGO-P150914",\n      SLACcitation   = "%%CITATION = ARXIV:1602.03837;%%"\n}\n'
+    assert inspire_citation(test_key) == known_output
+
+
+if __name__ == "__main__":
+
+    if len(sys.argv)==2:     # Get the name of the bibfile from the aux file
+        basename = sys.argv[1].split('.tex')[0]
[…]
