+++ example/example.tex
@@ -4,7 +4,7 @@
 
 \begin{document}
 
-\title{filltex test}
+\title{filltex example}
 
 \author{Davide Gerosa\\dgerosa@caltech.edu}
 
