### 2017-04-10 dgerosa/filltex#4 @dgerosa
Thanks for this comment.

- I have included two simple tests, called `test_ads` and `test_inspire`, to check the web scraping functions against known outputs. These are also described in the readme.
- The previous test.tex file was renamed example.tex
