### 2017-04-10 e016cc2
clean exmple files for new run
