+++ example/clean.sh
@@ -1,0 +1,1 @@
+find . -type f  ! -name "*.tex*" -delete
