### 2017-04-10 dgerosa/filltex#3 @dgerosa
I have implemented support for Python 3 in `fillbib`. Changes are:

- print statements
- import urllib.request instead of urllib
- changed filter(...) to list((filter(...))

The `filltex` script still calls `python`, and anything >=2.6 would now be supported. The user can change `python` to `python3`  in `filltext` to force python 3 usage.
