### 2017-04-10 dgerosa/filltex#6 @dgerosa
Thanks for all these ideas. I'll look into this soon and see if I want to rewrite the code as you suggest. In the meanwhile, any volunteer is more than welcome to give it a try!
