+++ README.md
@@ -126,4 +126,4 @@ The code is developed and maintained by [Davide Gerosa](www.davidegerosa.com). I
 
     dgerosa@caltech.edu
 
-The idea started from [this](http://www.vallis.org/salon/) `python` course taught by [Michele Vallisneri](http://www.vallis.org/) at Caltech (and in particular from [this example](http://www.vallis.org/salon/summary-2.html)). The [TexCount](http://app.uio.no/ifi/texcount) code is developed by Einar Andreas Rodland. [TexCount](http://app.uio.no/ifi/texcount) may be already included in your Tex distribution; in case it's not, you can find it in this repository for your convenience. Useful info on the INSPIRE and ADS APIs are available [here](https://inspirehep.net/info/hep/pub_list) and [here](https://github.com/adsabs/adsabs-dev-api).
+The idea started from [this](http://www.vallis.org/salon/) `python` course taught by [Michele Vallisneri](http://www.vallis.org/) at Caltech (and in particular from [this example](http://www.vallis.org/salon/summary-2.html)). We thank [Lars Holm Nielsen](@lnielsen), reviewer for [The Journal of Open Software](http://joss.theoj.org/), for several suggestions which improved `filltex`. The [TexCount](http://app.uio.no/ifi/texcount) code is developed by Einar Andreas Rodland. [TexCount](http://app.uio.no/ifi/texcount) may be already included in your Tex distribution; in case it's not, you can find it in this repository for your convenience. Useful info on the INSPIRE and ADS APIs are available [here](https://inspirehep.net/info/hep/pub_list) and [here](https://github.com/adsabs/adsabs-dev-api).
