### 2017-04-10 dgerosa/filltex#2 @lnielsen
If you include ``\cite{10.1051/0004-6361/201322068}`` in  ``example.tex`` you'll end up with ``@ARTICLE{2013A&A...558A..33A,…`` in the bib file. I.e. when you use a personal citation key and this key yield valid results in either INSPIRE/ADS you will include the entry in the bibliography which I don't think is desirable.

ADS bibcodes are [well-described](http://adsabs.harvard.edu/abs_doc/help_pages/data.html) and very simple to validate with a simple regular expression, and adding it will make the program more robust.

The INSPIRE citation keys are unfortunately not something I see that they have a standard for (perhaps @kaplun can correct me if I'm wrong), but you can e.g. make a search query that only searches the citation keys like this:

``035__a:Rowson:1985xh AND (035__9:SPIRESTeX OR 035__9:INSPIRETeX)``

The 035__a/9 is the MARC21 (a metadata standard for libraries) fields where the citation keys are stored. The ADS query can perhaps be fixed in a similar manner, but even though the query look likes it only query the bibcode it still give results if you give it e.g. DOI.

In the end these changes are just to make the program more robust when you have personal keys which I think is very likely in real world usage of filltex.

As for arXiv, I don't think it's the programs job to enforce a specific pattern on what an author should cite - i.e. an author can have fully valid reasons why they want to cite an arxiv paper. This doesn't mean you need to add support for arxiv, it just mean I wouldn't let it be your personal opinion on what people should cite be the deciding factor in this case.
