### 2017-04-10 dgerosa/filltex#2 @dgerosa
Thanks for spotting this issue with the DOI! I agree is not desirable: I didn't realize a query with the DOI on ADS would result in an item being downloaded.

The possible issue I see here is that DOIs are part of both ADS and INSPIRE; I would not like `filltex` to decide for the user which entry should be used. I'd rather leave DOIs entirely as personal keys, which are not filled with any database information.

I've placed simple if statements to check that the entry just downloaded is actually one that was originally requested. Including  a DOI key, like `\cite{10.1051/0004-6361/201322068}`, now does not write anything to the `.bib` file, as you were looking for`10.1051/0004-6361/201322068` but retrieved `2013A&A...558A..33A`.

This additional check does not apply to references cited with ADS preprint keys because in that case the input key and the retrieved key are different. Handling of these cases constitutes a known limitation of the code as mentioned in the readme, and should therefore be used with care.

Does this look acceptable to you? Thanks for all these comments, they are greatly improving this
 code.
