### 2017-04-10 66ec804
additional check in ads_citation and inspire_citation
