### 2017-04-21 dgerosa/filltex#1 @vallis
I've included the MIT license in the code at http://www.vallis.org/salon/summary-2.html. I'm very happy it's being used and improved here, and grant Davide all the rights required to do so.
