### 2017-04-21 #222 @dgerosa
Hello,

Michele @vallis and I addressed your comments. Please let us know if you find our modifications acceptable. Should I close the issues?

Best,
Davide
