### 2017-05-01 #222 @arfon
> Michele (@vallis) and I addressed your comments. Please let us know if you find our modifications acceptable. Should I close the issues?

@lnielsen - I think we should be good to accept here? Could you confirm?
