### 2017-05-02 #222 @arfon
> Yep, ok from my side. Sorry for the delay - was on extended vacation over easter :-)

👍 thanks for confirming @lnielsen

@dgerosa @vallis - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
