### 2017-05-02 #222 @dgerosa
Hello,

here is the badge for the new release (v1.1):
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.570875.svg)](https://doi.org/10.5281/zenodo.570875)

Best,
Davide
