### 2017-05-02 #222 @arfon
@dgerosa - apologies for being late to suggest this but could you move the references/links you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
