### 2017-05-02 60a5ba3
JOSS: links moved to .bib
