### 2017-05-02 #222 @dgerosa
Hello,

I don't have references. I moved the links into a .bib file, now in the JOSS directory.

Davide
