### 2017-05-02 #222 @arfon
Thanks @dgerosa. That's not quite it: you now need to cite these items in the `paper.bib` file (e.g. `@TexShop` - https://github.com/dgerosa/filltex/blob/master/JOSS/paper.bib#L25)
