### 2017-05-02 c20edbb
fixing citations
