### 2017-05-02 875b3f3
Update paper.md
