### 2017-05-02 #222 @arfon
@lnielsen - many thanks for the review here ✨

@dgerosa - your submission is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00222 :zap: 🚀 💥
